# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * League bowling app API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface User {
  id: number;
  username: string;
  name: string;
  rank: string;
  level: number;
  xp: number;
  xpToNext: number;
  isPro: boolean;
  careerAvg: number;
  highGame: number;
  totalGames: number;
  team: string;
  rating: number;
  createdAt: string;
  updatedAt: string;
}

export interface UpdateUserBody {
  name?: string;
  username?: string;
  team?: string;
  isPro?: boolean;
}

export interface Game {
  id: number;
  userId: number;
  score: number;
  date: string;
  alley: string;
  oilPattern: string;
  ballUsed: string;
  notes: string;
  verified: boolean;
  createdAt: string;
}

export interface CreateGameBody {
  score: number;
  date: string;
  alley: string;
  oilPattern: string;
  ballUsed: string;
  notes?: string;
  verified?: boolean;
}

export interface Challenge {
  id: number;
  userId: number;
  username: string;
  rank: string;
  rankColor: string;
  postedScore: number;
  stake: number;
  status: string;
  initials: string;
  avatarColor: string;
  isPro: boolean;
  timeAgo: string;
  createdAt: string;
}

export interface CreateChallengeBody {
  postedScore: number;
  stake: number;
}

export interface Moment {
  id: number;
  userId: number;
  username: string;
  rank: string;
  rankColor: string;
  content: string;
  /** @nullable */
  score?: number | null;
  type: string;
  likes: number;
  comments: number;
  initials: string;
  avatarColor: string;
  liked: boolean;
  timeAgo: string;
  createdAt: string;
}

export interface CreateMomentBody {
  content: string;
  /** @nullable */
  score?: number | null;
  type: string;
}

export interface League {
  id: number;
  name: string;
  description: string;
  members: number;
  type: string;
  level: string;
  avgScore: number;
  /** @nullable */
  weeklyChallenge?: string | null;
  joined: boolean;
  createdAt: string;
}