# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * League bowling app API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get current user profile
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  name: zod.string(),
  rank: zod.string(),
  level: zod.number(),
  xp: zod.number(),
  xpToNext: zod.number(),
  isPro: zod.boolean(),
  careerAvg: zod.number(),
  highGame: zod.number(),
  totalGames: zod.number(),
  team: zod.string(),
  rating: zod.number(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update current user profile
 */
export const UpdateMeBody = zod.object({
  name: zod.string().optional(),
  username: zod.string().optional(),
  team: zod.string().optional(),
  isPro: zod.boolean().optional(),
});

export const UpdateMeResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  name: zod.string(),
  rank: zod.string(),
  level: zod.number(),
  xp: zod.number(),
  xpToNext: zod.number(),
  isPro: zod.boolean(),
  careerAvg: zod.number(),
  highGame: zod.number(),
  totalGames: zod.number(),
  team: zod.string(),
  rating: zod.number(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary List current user's logged games
 */
export const ListGamesResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  score: zod.number(),
  date: zod.string(),
  alley: zod.string(),
  oilPattern: zod.string(),
  ballUsed: zod.string(),
  notes: zod.string(),
  verified: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const ListGamesResponse = zod.array(ListGamesResponseItem);

/**
 * @summary Log a new game
 */
export const CreateGameBody = zod.object({
  score: zod.number(),
  date: zod.string(),
  alley: zod.string(),
  oilPattern: zod.string(),
  ballUsed: zod.string(),
  notes: zod.string().optional(),
  verified: zod.boolean().optional(),
});

/**
 * @summary List open challenges from other players
 */
export const ListChallengesResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  username: zod.string(),
  rank: zod.string(),
  rankColor: zod.string(),
  postedScore: zod.number(),
  stake: zod.number(),
  status: zod.string(),
  initials: zod.string(),
  avatarColor: zod.string(),
  isPro: zod.boolean(),
  timeAgo: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListChallengesResponse = zod.array(ListChallengesResponseItem);

/**
 * @summary Post a new challenge
 */
export const CreateChallengeBody = zod.object({
  postedScore: zod.number(),
  stake: zod.number(),
});

/**
 * @summary List current user's active challenges
 */
export const ListMyChallengesResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  username: zod.string(),
  rank: zod.string(),
  rankColor: zod.string(),
  postedScore: zod.number(),
  stake: zod.number(),
  status: zod.string(),
  initials: zod.string(),
  avatarColor: zod.string(),
  isPro: zod.boolean(),
  timeAgo: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListMyChallengesResponse = zod.array(ListMyChallengesResponseItem);

/**
 * @summary Accept an open challenge
 */
export const AcceptChallengeParams = zod.object({
  id: zod.coerce.number(),
});

export const AcceptChallengeResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  username: zod.string(),
  rank: zod.string(),
  rankColor: zod.string(),
  postedScore: zod.number(),
  stake: zod.number(),
  status: zod.string(),
  initials: zod.string(),
  avatarColor: zod.string(),
  isPro: zod.boolean(),
  timeAgo: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary List social moments feed
 */
export const ListMomentsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  username: zod.string(),
  rank: zod.string(),
  rankColor: zod.string(),
  content: zod.string(),
  score: zod.number().nullish(),
  type: zod.string(),
  likes: zod.number(),
  comments: zod.number(),
  initials: zod.string(),
  avatarColor: zod.string(),
  liked: zod.boolean(),
  timeAgo: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListMomentsResponse = zod.array(ListMomentsResponseItem);

/**
 * @summary Post a moment
 */
export const CreateMomentBody = zod.object({
  content: zod.string(),
  score: zod.number().nullish(),
  type: zod.string(),
});

/**
 * @summary Like a moment
 */
export const LikeMomentParams = zod.object({
  id: zod.coerce.number(),
});

export const LikeMomentResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  username: zod.string(),
  rank: zod.string(),
  rankColor: zod.string(),
  content: zod.string(),
  score: zod.number().nullish(),
  type: zod.string(),
  likes: zod.number(),
  comments: zod.number(),
  initials: zod.string(),
  avatarColor: zod.string(),
  liked: zod.boolean(),
  timeAgo: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Unlike a moment
 */
export const UnlikeMomentParams = zod.object({
  id: zod.coerce.number(),
});

export const UnlikeMomentResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  username: zod.string(),
  rank: zod.string(),
  rankColor: zod.string(),
  content: zod.string(),
  score: zod.number().nullish(),
  type: zod.string(),
  likes: zod.number(),
  comments: zod.number(),
  initials: zod.string(),
  avatarColor: zod.string(),
  liked: zod.boolean(),
  timeAgo: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary List all leagues
 */
export const ListLeaguesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  members: zod.number(),
  type: zod.string(),
  level: zod.string(),
  avgScore: zod.number(),
  weeklyChallenge: zod.string().nullish(),
  joined: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const ListLeaguesResponse = zod.array(ListLeaguesResponseItem);

/**
 * @summary Join or request to join a league
 */
export const JoinLeagueParams = zod.object({
  id: zod.coerce.number(),
});

export const JoinLeagueResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  members: zod.number(),
  type: zod.string(),
  level: zod.string(),
  avgScore: zod.number(),
  weeklyChallenge: zod.string().nullish(),
  joined: zod.boolean(),
  createdAt: zod.coerce.date(),
});